+++ src/lib.rs
@@ -88,8 +88,8 @@
 //!   [`Deserialize`] traits for both [`f16`] and [`bf16`].
 //!
 //! - **`num-traits`** — Adds support for the [`num-traits`] crate by implementing [`ToPrimitive`],
-//!   [`FromPrimitive`], [`AsPrimitive`], [`Num`], [`Float`], [`FloatCore`], and [`Bounded`] traits
-//!   for both [`f16`] and [`bf16`].
+//!   [`FromPrimitive`], [`ToBytes`], [`FromBytes`], [`AsPrimitive`], [`Num`], [`Float`],
+//!   [`FloatCore`], and [`Bounded`] traits for both [`f16`] and [`bf16`].
 //!
 //! - **`bytemuck`** — Adds support for the [`bytemuck`] crate by implementing [`Zeroable`] and
 //!   [`Pod`] traits for both [`f16`] and [`bf16`].
@@ -151,6 +151,8 @@
     doc = "
 [`ToPrimitive`]: https://docs.rs/num-traits/*/num_traits/cast/trait.ToPrimitive.html
 [`FromPrimitive`]: https://docs.rs/num-traits/*/num_traits/cast/trait.FromPrimitive.html
+[`ToBytes`]: https://docs.rs/num-traits/*/num_traits/ops/bytes/trait.ToBytes.html
+[`FromBytes`]: https://docs.rs/num-traits/*/num_traits/ops/bytes/trait.FromBytes.html
 [`AsPrimitive`]: https://docs.rs/num-traits/*/num_traits/cast/trait.AsPrimitive.html
 [`Num`]: https://docs.rs/num-traits/*/num_traits/trait.Num.html
 [`Float`]: https://docs.rs/num-traits/*/num_traits/float/trait.Float.html